自由帳ページを別ファイルに
