+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:soccer_app_flutter/pages/note_page.dart';
 import 'package:soccer_app_flutter/theme/button_theme_extension.dart';
 import 'package:soccer_app_flutter/utils/layout_helpers.dart';
 import 'package:soccer_app_flutter/widgets/circle_button.dart';
@@ -374,24 +375,6 @@ class ConnectionPageState extends State<ConnectionPage> {
             ],
           );
         },
-      ),
-    );
-  }
-}
-
-// NotePage（自由帳ページ）
-class NotePage extends StatelessWidget {
-  const NotePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        title: const Text('自由帳 デモページ'),
-      ),
-      body: const Center(
-        child: Text('このページに自由帳の内容を記載予定', style: TextStyle(fontSize: 24)),
       ),
     );
   }
